vz-4021 refactor framework to be under pkg/test
