+++ tests/e2e/verify-install/kubernetes/kubernetes_test.go
@@ -7,9 +7,10 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/verrazzano/verrazzano/pkg/test/framework"
+
 	ginkgoExt "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
-	"github.com/verrazzano/verrazzano/pkg/framework"
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	v1 "k8s.io/api/core/v1"
