+++ tests/e2e/verify-install/verrazzano/verrazzano_test.go
@@ -6,9 +6,10 @@ package verrazzano_test
 import (
 	"time"
 
+	"github.com/verrazzano/verrazzano/pkg/test/framework"
+
 	ginkgoExt "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
-	"github.com/verrazzano/verrazzano/pkg/framework"
 	"github.com/verrazzano/verrazzano/tests/e2e/pkg"
 	rbacv1 "k8s.io/api/rbac/v1"
 )
